huaweicloud_scm_certificate adds support for pushing certificates.(#1179)
